cn0414: add platform drivers for I2C, SPI, GPIO, UART and delay
Implement an abstraction layer between the platform drivers and
functions used by the application and device dirvers. These files
contain implementation for the I2C, SPI and UART communication achieved
with the microcontrollers hardware modules. They also include an
implementation for the GPIO controller and a generic milisecond delay
function.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>