cn0414: add software UART implementation
This application requires two UART lines: one for the user CLI and the
other for interfacing with the AD5700 HART modem. As the microcontroller
has only one UART module and the speed for the HART communication is
only 1200 baud rate, this software UART library is used to drive the
HART communication. It is done using GPIOs and timer 1 to create UART
messages by bitbanging at specific intervals.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>